+++ core/src/elements/ons-bottom-toolbar.spec.js
@@ -40,8 +40,8 @@ describe('ons-bottom-toolbar', () => {
 
   describe('#_compile()', () => {
     it('does not compile twice', () => {
-      let div1 = document.createElement('div');
-      let div2 = document.createElement('div');
+      const div1 = document.createElement('div');
+      const div2 = document.createElement('div');
       div1.innerHTML = '<ons-bottom-toolbar></ons-bottom-toolbar>';
       div2.innerHTML = div1.innerHTML;
       expect(div1.isEqualNode(div2)).to.be.true;