+++ core/src/elements/ons-list.spec.js
@@ -28,8 +28,8 @@ describe('ons-list', () => {
 
   describe('#_compile()', () => {
     it('does not compile twice', () => {
-      let div1 = document.createElement('div');
-      let div2 = document.createElement('div');
+      const div1 = document.createElement('div');
+      const div2 = document.createElement('div');
       div1.innerHTML = `
         <ons-list>
           <ons-list-header>Content</ons-list-header>
@@ -45,7 +45,7 @@ describe('ons-list', () => {
   describe('autoStyling', () => {
     it('adds \'material\' modifier on Android', () => {
       ons.platform.select('android');
-      let e = document.createElement('ons-list');
+      const e = document.createElement('ons-list');
       expect(e.getAttribute('modifier')).to.equal('material');
       ons.platform.select('');
     });