+++ core/src/elements/ons-navigator/index.js
@@ -356,7 +356,7 @@ class NavigatorElement extends BaseElement {
       options.animationOptions = util.extend({}, leavePage.pushedOptions.animationOptions, options.animationOptions || {});
 
       const callback = () => {
-        enterPage._show();
+        setImmediate(() => enterPage._show());
         leavePage._hide();
 
         pages.pop();
@@ -456,7 +456,7 @@ class NavigatorElement extends BaseElement {
       enterPage.updateBackButton(pageLength - 1);
 
       enterPage.data = options.data;
-      enterPage._name = options.page;
+      enterPage.name = options.page;
       enterPage.pushedOptions = options;
 
       return new Promise(resolve => {
@@ -482,10 +482,10 @@ class NavigatorElement extends BaseElement {
           enterPage.style.display = 'block';
           if (leavePage) {
             leavePage._hide();
-            enterPage._show();
+            setImmediate(() => enterPage._show());
             animator.push(enterPage, leavePage, done);
           } else {
-            enterPage._show();
+            setImmediate(() => enterPage._show());
             done();
           }
         };
@@ -543,7 +543,7 @@ class NavigatorElement extends BaseElement {
 
     const run = templateHTML => {
       const element = util.extend(this._createPageElement(templateHTML), {
-        _name: options.page,
+        name: options.page,
         data: options.data,
         pushedOptions: options
       });
@@ -818,14 +818,14 @@ class NavigatorElement extends BaseElement {
   }
 
   _show() {
-    if (this.pages[this.pages.length - 1]) {
-      this.pages[this.pages.length - 1]._show();
+    if (this.topPage) {
+      this.topPage._show();
     }
   }
 
   _hide() {
-    if (this.pages[this.pages.length - 1]) {
-      this.pages[this.pages.length - 1]._hide();
+    if (this.topPage) {
+      this.topPage._hide();
     }
   }