+++ core/src/elements/ons-speed-dial.js
@@ -22,25 +22,26 @@ const scheme = {
 
 /**
  * @element ons-speed-dial
- * @category speeddial
+ * @category speed-dial
  * @description
- *   [en]Element that displays a Material Design Speed Dialog component.[/en]
- *   [ja]Material DesignのSpeed dialコンポーネントを表現する要素です。[/ja]
+ *   [en]
+ *     Element that displays a Material Design Speed Dialog component. It is useful when there are more than one primary action that can be performed in a page.
+ *
+ *     The Speed dial looks like a `<ons-fab>` element but will expand a menu when tapped.
+ *   [/en]
+ *   [ja][/ja]
  * @codepen dYQYLg
  * @seealso ons-speed-dial-item
- *   [en]ons-speed-dial-item component[/en]
+ *   [en]The `<ons-speed-dial-item>` represents a menu item.[/en]
  *   [ja]ons-speed-dial-itemコンポーネント[/ja]
  * @example
  * <ons-speed-dial position="left bottom">
- *   <ons-icon
- *     icon="fa-twitter"
- *     size="26px"
- *     fixed-width="false"
- *     style="vertical-align:middle;">
- *   </ons-icon>
- *   <ons-speed-dial-item><ons-ripple></ons-ripple>C</ons-speed-dial-item>
- *   <ons-speed-dial-item><ons-ripple></ons-ripple>B</ons-speed-dial-item>
- *   <ons-speed-dial-item><ons-ripple></ons-ripple>A</ons-speed-dial-item>
+ *   <ons-fab>
+ *     <ons-icon icon="fa-twitter"></ons-icon>
+ *   </ons-fab>
+ *   <ons-speed-dial-item>A</ons-speed-dial-item>
+ *   <ons-speed-dial-item>B</ons-speed-dial-item>
+ *   <ons-speed-dial-item>C</ons-speed-dial-item>
  * </ons-speed-dial>
  */
 class SpeedDialElement extends BaseElement {
@@ -133,6 +134,7 @@ class SpeedDialElement extends BaseElement {
   }
 
   attachedCallback() {
+    this._updateDirection(this.hasAttribute('direction') ? this.getAttribute('direction') : 'up');
     this.addEventListener('click', this._boundOnClick, false);
   }
 
@@ -164,7 +166,10 @@ class SpeedDialElement extends BaseElement {
 
   _updateRipple() {
     const fab = util.findChild(this, 'ons-fab');
-    this.hasAttribute('ripple') ? fab.setAttribute('ripple', '') : fab.removeAttribute('ripple');
+
+    if (fab) {
+      this.hasAttribute('ripple') ? fab.setAttribute('ripple', '') : fab.removeAttribute('ripple');
+    }
   }
 
   _updateDirection(direction) {
@@ -317,15 +322,15 @@ class SpeedDialElement extends BaseElement {
    * @property disabled
    * @type {Boolean}
    * @description
-   *   [en]Whether the speed dial is disabled or not.[/en]
-   *   [ja][/ja]
+   *   [en]Whether the element is disabled or not.[/en]
+   *   [ja]無効化されている場合に`true`。[/ja]
    */
   set disabled(value) {
     if (value) {
       this.hideItems();
     }
     util.arrayFrom(this.children).forEach(e => {
-      util.match(e, '.fab') && util.toggleAttribute(e, 'disabled', value)
+      util.match(e, '.fab') && util.toggleAttribute(e, 'disabled', value);
     });
 
     return util.toggleAttribute(this, 'disabled', value);
@@ -340,8 +345,8 @@ class SpeedDialElement extends BaseElement {
    * @readonly
    * @type {Boolean}
    * @description
-   *   [en]Whether the component is inline or not.[/en]
-   *   [ja]この要素がインライン要素かどうか。[/ja]
+   *   [en]Whether the element is inline or not.[/en]
+   *   [ja]インライン要素の場合に`true`。[/ja]
    */
   get inline() {
     return this.hasAttribute('inline');
@@ -352,8 +357,8 @@ class SpeedDialElement extends BaseElement {
    * @readonly
    * @type {Boolean}
    * @description
-   *   [en]Whether the dialog is visible or not.[/en]
-   *   [ja]ダイアログが表示されているかどうか。[/ja]
+   *   [en]Whether the element is visible or not.[/en]
+   *   [ja]要素が見える場合に`true`。[/ja]
    */
   get visible() {
     return this._shown && this.style.display !== 'none';